+++ common/sequencer_timing.cpp
@@ -744,7 +744,7 @@ static void prepare_signature_and_tempo_automation(QVector<SignatureAutomationNo
       break;
 
 #if !defined(RELEASE)
-    printf("      I: %d. is_changed: %d. max1: %d. max2: %d. size1: %d. size2: %d\n", i, is_changed, max_size1, max_size2, signatures.size(), tempos.size());
+    printf("      I: %d. is_changed: %d. max1: %d. max2: %d. size1: %d. size2: %d\n", i, is_changed, max_size1, max_size2, (int)signatures.size(), (int)tempos.size());
     
     if (i > R_MAX(init_signature_size, init_tempos_size)){
       printf("  Signatures for #%d:\n", i);
@@ -759,7 +759,7 @@ static void prepare_signature_and_tempo_automation(QVector<SignatureAutomationNo
 #endif
     
     if (i > max_size1){
-      printf("Warning:      I: %d. is_changed: %d. max1: %d. max2: %d. size1: %d. size2: %d\n", i, is_changed, max_size1, max_size2, signatures.size(), tempos.size());
+		printf("Warning:      I: %d. is_changed: %d. max1: %d. max2: %d. size1: %d. size2: %d\n", i, is_changed, max_size1, max_size2, (int)signatures.size(), (int)tempos.size());
       //getchar();
       R_ASSERT_NON_RELEASE(false);
     }
